TCPSocket.hh - bug in receive() fixed (result of recv was caught by size_t but should have been signed integer)

+++ includes/CxxUtilities/TCPSocket.hh
@@ -195,7 +195,7 @@ public:
 	 * @return received size
 	 */
 	size_t receive(uint8_t* data, uint32_t length, bool waitUntilSpecifiedLengthCompletes = false) throw (TCPSocketException) {
-		size_t result = 0;
+		int result = 0;
 		int remainingLength=length;
 		int readDoneLength=0;
